requirements.txt
streamlit==1.46.0
pandas==1.5.3
pandasai==2.0.44
pymongo==4.13.2
python-dotenv==1.1.0
werkzeug==3.1.3
Pillow==10.4.0
openai==1.90.0